Reorganize some of the cpp
