+++ R/RcppExports.R
@@ -1,12 +1,8 @@
 # Generated by using Rcpp::compileAttributes() -> do not edit by hand
 # Generator token: 10BE3573-1514-4C36-9D1C-5A225CD40393
 
-point_mult <- function(X, Y) {
-    invisible(.Call('_FastProjectR_point_mult', PACKAGE = 'FastProjectR', X, Y))
 }
 
-multMat <- function(X, Y) {
-    .Call('_FastProjectR_multMat', PACKAGE = 'FastProjectR', X, Y)
 }
 
 ball_tree_vector_knn <- function(X, Y, K, n_threads) {
@@ -19,5 +15,13 @@ ball_tree_knn <- function(X, K, n_threads) {
 
 load_in_knn <- function(nn, d) {
     .Call('_FastProjectR_load_in_knn', PACKAGE = 'FastProjectR', nn, d)
+}
+
+point_mult <- function(X, Y) {
+    invisible(.Call('_FastProjectR_point_mult', PACKAGE = 'FastProjectR', X, Y))
+}
+
+multMat <- function(X, Y) {
+    .Call('_FastProjectR_multMat', PACKAGE = 'FastProjectR', X, Y)
 }
